split AggregateEvaluators to clean this file (closes #90)
