+++ index.ts
@@ -1,6 +1,7 @@
 export { AsyncEvaluator, AsyncEvaluatorConfig } from './lib/evaluators/AsyncEvaluator';
 export { SyncEvaluator, SyncEvaluatorConfig } from './lib/evaluators/SyncEvaluator';
-export { AggregateEvaluator, AsyncAggregateEvaluator } from './lib/evaluators/AggregateEvaluator';
+export { AggregateEvaluator} from './lib/evaluators/AggregateEvaluator';
 
 export { ExpressionError, isExpressionError } from './lib/util/Errors';
 export { orderTypes } from './lib/util/Ordering';
+export {AsyncAggregateEvaluator} from './lib/evaluators/AsyncAggregateEvaluator';
